TO DO: conversion to float16 for wmma

+++ flash_attention_tensor_cores/testing.cu
@@ -131,9 +131,9 @@ int main(){
     const int number_of_warps = 4;
     const int tile = 16 * 4;
     int shared_mem_needed = D * tile * sizeof(__half); // queries
-    shared_mem_needed += D * tile * sizeof(__half); // output
+    shared_mem_needed += D * tile * sizeof(float); // output
     shared_mem_needed += D * tile * sizeof(__half); // keys + values idepdendently
-    shared_mem_needed += tile * tile * sizeof(__half); // tile
+    shared_mem_needed += tile * tile * sizeof(float); // tile
     shared_mem_needed += 64 * sizeof(__half); // running max
     shared_mem_needed += 64 * sizeof(float); // running sum
